Propose WebM first or else the browser will not use it
Browsers don't speculate about which `<source>` is optimal, so the order
of `<source>`s matters. If you specify an MPEG-4 video first and
the browser supports WebM, browsers will skip the WebM `<source>` and use the
MPEG-4 instead. If you prefer a WebM `<source>` be used first, _specify it
first!_